api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
beautifulsoup4==4.13.5
    # via yfinance
blinker==1.9.0
    # via flask
certifi==2025.8.3
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via curl-cffi
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   flask
    #   uvicorn
curl-cffi==0.13.0
    # via yfinance
exceptiongroup==1.3.0
    # via anyio
fastapi==0.115.0
    # via -r requirements.in
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==4.0.2
    # via -r requirements.in
frozendict==2.4.6
    # via yfinance
gunicorn==21.2.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.27.0
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
multitasking==0.0.12
    # via yfinance
numpy==2.1.0
    # via
    #   -r requirements.in
    #   pandas
    #   yfinance
packaging==25.0
    # via gunicorn
pandas==2.2.2
    # via
    #   -r requirements.in
    #   yfinance
peewee==3.18.2
    # via yfinance
platformdirs==4.4.0
    # via yfinance
protobuf==6.32.0
    # via yfinance
psutil==6.0.0
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.20.1
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via uvicorn
pytz==2025.2
    # via
    #   pandas
    #   yfinance
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via
    #   -r requirements.in
    #   yfinance
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
soupsieve==2.8
    # via beautifulsoup4
starlette==0.38.6
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   beautifulsoup4
    #   exceptiongroup
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   uvicorn
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.30.1
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via
    #   uvicorn
    #   yfinance
werkzeug==3.1.3
    # via flask
yfinance==0.2.65
    # via -r requirements.in
